+++ buession-springboot-cache/src/main/java/com/buession/springboot/cache/redis/autoconfigure/RedisConfiguration.java
@@ -27,8 +27,6 @@ package com.buession.springboot.cache.redis.autoconfigure;
 import com.buession.redis.RedisTemplate;
 import com.buession.redis.client.connection.datasource.DataSource;
 import com.buession.redis.core.Options;
-import com.buession.springboot.cache.redis.core.RedisCacheInvocationHandler;
-import com.buession.springboot.cache.redis.core.RedisCacheManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
@@ -37,8 +35,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import java.lang.reflect.Proxy;
 
 /**
  * Redis 自动配置类
@@ -75,18 +71,6 @@ public class RedisConfiguration {
 		}
 
 		return template;
-	}
-
-	@Bean
-	@ConditionalOnClass(name = {"com.buession.aop.DefaultMethodInvoker"})
-	@ConditionalOnBean(RedisTemplate.class)
-	public RedisCacheManager redisCacheManager(RedisTemplate redisTemplate) {
-		final Class<?>[] interfaces = new Class[]{RedisCacheManager.class};
-
-		final RedisCacheInvocationHandler cacheInvocationHandler = new RedisCacheInvocationHandler(redisTemplate,
-				interfaces);
-		return (RedisCacheManager) Proxy.newProxyInstance(RedisTemplate.class.getClassLoader(), interfaces,
-				cacheInvocationHandler);
 	}
 
 }